+++ src/Resume.tsx
@@ -2,7 +2,7 @@ import { Certificates } from "./components/Certificates";
 import { Education } from "./components/Education";
 import { Employment } from "./components/Employment";
 import { Interests } from "./components/Interests";
-import { PersonalDetails } from "./components/PersonalDetails";
+import { PersonalDetails } from "./components/personal-details/personal-details";
 import { Qualifications } from "./components/Qualifications";
 import { References } from "./components/References";
 import { Skills } from "./components/Skills";
